Add data for k=1 and k=2 (that was run correctly this time)
There are multiple files for k=1 because there were errors in the log. It's not connectivity problems but rather is that FINVALUE RPCs were failing. I think this is because with k=1, the buckets are too small to hold enough information to always route to the proper location. I will verify with the paper. We can still use the data we have available and then also mention it during class.